- refs #81965: Moved decimal places field to the last step.
